+++ specstory-cli/changelog.md
@@ -4,7 +4,8 @@
 
 ### ⚙️ Improvements
 
-- New `specstory list` command to list all sessions for all supported providers for the current directory. Can also be used to list sessions for a specific provider, e.g. `specstory list claude`.
+- New `specstory list` command to list all sessions for all supported providers for the current project directory. Can also be used to list sessions for a specific provider, e.g. `specstory list claude`.
+- `specstory list --json` flag for easily parsable output.
 
 ## v1.5.0 2026-01-31